+++ local-notifications/android/src/main/java/com/capacitorjs/plugins/localnotifications/LocalNotificationRestoreReceiver.java
@@ -1,11 +1,8 @@
 package com.capacitorjs.plugins.localnotifications;
 
-import static android.os.Build.VERSION.SDK_INT;
-
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.os.Build;
 import android.os.UserManager;
 import com.getcapacitor.CapConfig;
 import java.util.ArrayList;
@@ -16,10 +13,8 @@ public class LocalNotificationRestoreReceiver extends BroadcastReceiver {
 
     @Override
     public void onReceive(Context context, Intent intent) {
-        if (SDK_INT >= Build.VERSION_CODES.N) {
-            UserManager um = context.getSystemService(UserManager.class);
-            if (um == null || !um.isUserUnlocked()) return;
-        }
+        UserManager um = context.getSystemService(UserManager.class);
+        if (um == null || !um.isUserUnlocked()) return;
 
         NotificationStorage storage = new NotificationStorage(context);
         List<String> ids = storage.getSavedNotificationIds();