+++ local-notifications/src/definitions.ts
@@ -35,11 +35,11 @@ declare module '@capacitor/cli' {
       /**
        * Set the default notification sound for notifications.
        *
-       * On Android 26+ it sets the default channel sound and can't be
+       * On Android 8+ it sets the default channel sound and can't be
        * changed unless the app is uninstalled.
        *
        * If the audio file is not found, it will result in the default system
-       * sound being played on Android 21-25 and no sound on Android 26+.
+       * sound being played on Android 7.x and no sound on Android 8+.
        *
        * Only available for Android.
        *
@@ -103,9 +103,7 @@ export interface LocalNotificationsPlugin {
    *
    * @since 4.0.0
    */
-  removeDeliveredNotifications(
-    delivered: DeliveredNotifications,
-  ): Promise<void>;
+  removeDeliveredNotifications(delivered: DeliveredNotifications): Promise<void>;
 
   /**
    * Remove all the notifications from the notifications screen.
@@ -591,8 +589,8 @@ export interface LocalNotificationSchema {
    *
    * Recommended format is `.wav` because is supported by both iOS and Android.
    *
-   * Only available for iOS and Android < 26.
-   * For Android 26+ use channelId of a channel configured with the desired sound.
+   * Only available for iOS and Android 7.x.
+   * For Android 8+ use channelId of a channel configured with the desired sound.
    *
    * If the sound file is not found, (i.e. empty string or wrong name)
    * the default system notification sound will be used.
@@ -722,7 +720,7 @@ export interface LocalNotificationSchema {
    * [`NotificationCompat.Builder`](https://developer.android.com/reference/androidx/core/app/NotificationCompat.Builder)
    * with the provided value.
    *
-   * Only available for Android 26+.
+   * Only available for Android 8+.
    *
    * @since 1.0.0
    */
@@ -803,8 +801,6 @@ export interface Schedule {
   /**
    * Allow this notification to fire while in [Doze](https://developer.android.com/training/monitoring-device-state/doze-standby)
    *
-   * Only available for Android 23+.
-   *
    * Note that these notifications can only fire [once per 9 minutes, per app](https://developer.android.com/training/monitoring-device-state/doze-standby#assessing_your_app).
    *
    * @since 1.0.0
@@ -849,15 +845,7 @@ export interface ScheduleOn {
   second?: number;
 }
 
-export type ScheduleEvery =
-  | 'year'
-  | 'month'
-  | 'two-weeks'
-  | 'week'
-  | 'day'
-  | 'hour'
-  | 'minute'
-  | 'second';
+export type ScheduleEvery = 'year' | 'month' | 'two-weeks' | 'week' | 'day' | 'hour' | 'minute' | 'second';
 
 export interface ListChannelsResult {
   /**