+++ status-bar/Package.swift
@@ -4,14 +4,14 @@ import PackageDescription
 
 let package = Package(
     name: "CapacitorStatusBar",
-    platforms: [.iOS(.v14)],
+    platforms: [.iOS(.v15)],
     products: [
         .library(
             name: "CapacitorStatusBar",
             targets: ["StatusBarPlugin"])
     ],
     dependencies: [
-        .package(url: "https://github.com/ionic-team/capacitor-swift-pm.git", from: "7.0.0")
+        .package(url: "https://github.com/ionic-team/capacitor-swift-pm.git", from: "8.0.0-beta")
     ],
     targets: [
         .target(