+++ toast/Package.swift
@@ -3,14 +3,14 @@ import PackageDescription
 
 let package = Package(
     name: "CapacitorToast",
-    platforms: [.iOS(.v14)],
+    platforms: [.iOS(.v15)],
     products: [
         .library(
             name: "CapacitorToast",
             targets: ["ToastPlugin"])
     ],
     dependencies: [
-        .package(url: "https://github.com/ionic-team/capacitor-swift-pm.git", from: "7.0.0")
+        .package(url: "https://github.com/ionic-team/capacitor-swift-pm.git", from: "8.0.0-beta")
     ],
     targets: [
         .target(