+++ toast/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.0.0-beta.0](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/toast@8.0.0-alpha.1...@capacitor/toast@8.0.0-beta.0) (2025-11-14)
+
+**Note:** Version bump only for package @capacitor/toast
+
+# [8.0.0-alpha.1](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/toast@7.0.2...@capacitor/toast@8.0.0-alpha.1) (2025-09-08)
+
+**Note:** Version bump only for package @capacitor/toast
+
+## [7.0.2](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/toast@7.0.1...@capacitor/toast@7.0.2) (2025-08-05)
+
+**Note:** Version bump only for package @capacitor/toast
+
 ## [7.0.1](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/toast@7.0.0...@capacitor/toast@7.0.1) (2025-04-02)
 
 **Note:** Version bump only for package @capacitor/toast